register bug, image for messages

+++ FrontEnd/app/(tabs)/(profiles)/settings/account.jsx
@@ -36,7 +36,9 @@ export default function AccountSettings() {
                         text: "Log out",
                         onPress: async () => {
                             await asyncStorage.clear();
-                            localStorage.clear();
+                            if (Platform.OS === "web") {
+                                localStorage.clear();
+                            }
                             if (Platform.OS !== "web") {
                                 await SecureStore.deleteItemAsync("token");
                                 await SecureStore.deleteItemAsync("username");
